Add a game to the db

+++ Web/DamaGame.Web/Controllers/PlayersController.cs
@@ -1,11 +1,13 @@
 ﻿namespace DamaGame.Web.Controllers
 {
+    using System.Diagnostics;
     using System.Linq;
     using System.Threading.Tasks;
 
     using DamaGame.Data.Common.Repositories;
     using DamaGame.Data.Models;
     using DamaGame.Services.Data;
+    using DamaGame.Web.ViewModels;
     using DamaGame.Web.ViewModels.Players;
     using Microsoft.AspNetCore.Identity;
     using Microsoft.AspNetCore.Mvc;
@@ -36,6 +38,13 @@
         [HttpPost]
         public async Task<IActionResult> AddPlayerAsync(PlayerInputViewModel input)
         {
+            if (this.playersRepository.All().Any(x => x.Name == input.Name)
+                || input.Name.Length < 4
+                || input.Name.Length > 16)
+            {
+                return this.RedirectToAction("AddPlayerError", "Errors");
+            }
+
             var user = await this.userManager.GetUserAsync(this.User);
 
             await this.playersService.InsertPlayer(input, user);
@@ -51,6 +60,13 @@
         [HttpPost]
         public async Task<IActionResult> RemovePlayerAsync(string name)
         {
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            if (!this.playersRepository.All().Any(x => x.Name == name && x.User == user))
+            {
+                return this.RedirectToAction("RemovePlayerError", "Errors");
+            }
+
             await this.playersService.RemovePlayer(name);
 
             return this.RedirectToAction("Beginning", "Beginning");
